+++ Accessibility-Handbook-Info.plist
@@ -6,7 +6,11 @@
 	<array>
 		<dict>
 			<key>CFBundleTypeRole</key>
-			<string>Editor</string>
+			<string>Viewer</string>
+			<key>CFBundleURLIconFile</key>
+			<string></string>
+			<key>CFBundleURLName</key>
+			<string>pereira.nascimento.giovani.accHandbook</string>
 			<key>CFBundleURLSchemes</key>
 			<array>
 				<string>accessibility_handbook</string>
